[component/client] Move PostBox and PostList to one file
- So far so good; next up - splitting them
